refactor: split App Source into smaller components
