--- app/common/renderer/components/SessionInspector/SourceTab/AppSource.jsx
@@ -1,272 +1,0 @@
-import {CaretDownOutlined} from '@ant-design/icons';
-import {
-  IconDownload,
-  IconEyeCode,
-  IconFiles,
-  IconFileText,
-  IconFold,
-  IconSearch,
-} from '@tabler/icons-react';
-import {Button, Card, Flex, Input, Row, Space, Spin, Tooltip, Tree} from 'antd';
-import {useCallback, useMemo, useState} from 'react';
-import {useTranslation} from 'react-i18next';
-
-import {BUTTON, ROW} from '../../../constants/antd-types.js';
-import {IMPORTANT_SOURCE_ATTRS} from '../../../constants/source.js';
-import {downloadFile} from '../../../utils/file-handling.js';
-import {copyToClipboard} from '../../../utils/other.js';
-import inspectorStyles from '../SessionInspector.module.css';
-import styles from './Source.module.css';
-
-const downloadXML = (sourceXML) => {
-  const href = 'data:application/xml;charset=utf-8,' + encodeURIComponent(sourceXML);
-  const filename = `app-source-${new Date().toJSON()}.xml`;
-  downloadFile(href, filename);
-};
-
-/**
- * Shows the 'source' of the app as a Tree
- */
-const AppSource = (props) => {
-  const {
-    sourceJSON,
-    sourceError,
-    setExpandedPaths,
-    expandedPaths,
-    selectedElement = {},
-    showSourceAttrs,
-    methodCallInProgress,
-    isUsingMjpegMode,
-    isSourceRefreshOn,
-    sourceXML,
-    toggleShowAttributes,
-  } = props;
-  const {t} = useTranslation();
-
-  const [searchValue, setSearchValue] = useState('');
-  const [autoExpandParent, setAutoExpandParent] = useState(true);
-
-  const getHighlightedText = (text) => {
-    if (!searchValue) {
-      return text;
-    }
-
-    const index = text.toLowerCase().indexOf(searchValue.toLowerCase());
-    if (index < 0) {
-      return text;
-    }
-
-    const prefix = text.substring(0, index);
-    const suffix = text.slice(index + searchValue.length);
-    const matchedWord = text.slice(index, index + searchValue.length);
-
-    return (
-      <>
-        {prefix}
-        <span className={styles.sourceSearchHighlight}>{matchedWord}</span>
-        {suffix}
-      </>
-    );
-  };
-
-  const getFormattedTag = (el, showAllAttrs) => {
-    const {tagName, attributes} = el;
-    let attrs = [];
-
-    for (let attr of Object.keys(attributes)) {
-      if ((IMPORTANT_SOURCE_ATTRS.includes(attr) && attributes[attr]) || showAllAttrs) {
-        attrs.push(
-          <span key={attr}>
-            &nbsp;
-            <span className={styles.sourceNodeAttrName}>{getHighlightedText(attr)}</span>=&quot;
-            <span className={styles.sourceNodeAttrValue}>
-              {getHighlightedText(attributes[attr])}
-            </span>
-            &quot;
-          </span>,
-        );
-      }
-    }
-    return (
-      <span className={inspectorStyles.monoFont}>
-        &lt;<span className={styles.sourceNodeTag}>{getHighlightedText(tagName)}</span>
-        {attrs}&gt;
-      </span>
-    );
-  };
-
-  /**
-   * Binds to antd Tree onSelect. If an item is being unselected, path is undefined
-   * otherwise 'path' refers to the element's path.
-   */
-  const handleSelectElement = (path) => {
-    const {selectElement, unselectElement} = props;
-
-    if (!path) {
-      unselectElement();
-    } else {
-      selectElement(path);
-    }
-  };
-
-  // Recursives through the source and renders a TreeNode for an element
-  const recursive = (elemObj) => {
-    if (!((elemObj || {}).children || []).length) {
-      return null;
-    }
-
-    return elemObj.children.map((el) => ({
-      title: getFormattedTag(el, showSourceAttrs),
-      key: el.path,
-      children: recursive(el),
-    }));
-  };
-
-  const treeData = sourceJSON && recursive(sourceJSON);
-
-  const flatten = (elemObj) => [elemObj, ...(elemObj.children?.flatMap(flatten) || [])];
-
-  const flatTreeData = sourceJSON && flatten(sourceJSON);
-
-  const elementMatchesSearch = useCallback(
-    (element, searchValue) => {
-      const checkedAttrTexts = Object.entries(element.attributes)
-        .filter(([name]) => IMPORTANT_SOURCE_ATTRS.includes(name) || showSourceAttrs)
-        .map(([name, value]) => name + value);
-      const allCheckedTexts = element.tagName + checkedAttrTexts;
-      return allCheckedTexts.toLowerCase().includes(searchValue.toLowerCase());
-    },
-    [showSourceAttrs],
-  );
-
-  const matchingElements = useMemo(
-    () =>
-      searchValue && flatTreeData
-        ? flatTreeData.filter((el) => elementMatchesSearch(el, searchValue))
-        : [],
-    [searchValue, flatTreeData, elementMatchesSearch],
-  );
-
-  // No need to recalculate if e.g. attribute visibility is toggled
-  const expandedKeys = useMemo(
-    () => [...matchingElements.map((el) => el.path), ...expandedPaths],
-    [matchingElements, expandedPaths],
-  );
-
-  const onChange = (e) => {
-    const {value} = e.target;
-    setSearchValue(value);
-    setAutoExpandParent(value !== '');
-  };
-
-  const onExpand = (expandedPaths) => {
-    setExpandedPaths(expandedPaths);
-    setAutoExpandParent(false);
-  };
-
-  const collapseAll = () => {
-    setExpandedPaths([]);
-    setAutoExpandParent(false);
-  };
-
-  return (
-    <Card
-      title={
-        <Flex gap={4} align="center">
-          <IconFileText size={18} />
-          {t('App Source')}
-        </Flex>
-      }
-      extra={
-        <span>
-          <Tooltip title={t('Copy XML Source to Clipboard')}>
-            <Button
-              type="text"
-              id="btnSourceXML"
-              icon={<IconFiles size={18} />}
-              onClick={() => copyToClipboard(sourceXML)}
-            />
-          </Tooltip>
-          <Tooltip title={t('Download Source as .XML File')}>
-            <Button
-              type="text"
-              id="btnDownloadSourceXML"
-              icon={<IconDownload size={18} />}
-              onClick={() => downloadXML(sourceXML)}
-            />
-          </Tooltip>
-        </span>
-      }
-    >
-      <div id="sourceContainer" className={styles.treeContainer} tabIndex="0">
-        {!sourceJSON && !sourceError && <i>{t('Gathering initial app source…')}</i>}
-        {sourceError && t('couldNotObtainSource', {errorMsg: JSON.stringify(sourceError)})}
-        {/* Show loading indicator in MJPEG mode if a method call is in progress and source refresh is on */}
-        <Spin
-          size="large"
-          spinning={!!methodCallInProgress && isUsingMjpegMode && isSourceRefreshOn}
-        >
-          {/* Must switch to a new antd Tree component when there's changes to treeData  */}
-          {treeData ? (
-            <div className={styles.treeWrapper}>
-              <Row justify="center" type={ROW.FLEX} align="middle" className={styles.treeActions}>
-                <Space.Compact>
-                  <Tooltip title={t('Collapse All')}>
-                    <Button
-                      id="btnCollapseAll"
-                      icon={<IconFold size={18} />}
-                      onClick={collapseAll}
-                    />
-                  </Tooltip>
-                  <Tooltip title={t('Toggle Attributes')}>
-                    <Button
-                      id="btnToggleAttrs"
-                      icon={<IconEyeCode size={18} />}
-                      onClick={toggleShowAttributes}
-                      type={showSourceAttrs ? BUTTON.PRIMARY : BUTTON.DEFAULT}
-                    />
-                  </Tooltip>
-                </Space.Compact>
-                <Space.Compact className={styles.treeSearchInput}>
-                  <Input
-                    placeholder={t('Search Source')}
-                    onChange={onChange}
-                    value={searchValue}
-                    allowClear
-                    prefix={<IconSearch size={12} />}
-                  />
-                  <Space.Addon className={styles.treeSearchInputAddon}>
-                    <Tooltip title={t('Matching Elements')}>{matchingElements.length}</Tooltip>
-                  </Space.Addon>
-                </Space.Compact>
-              </Row>
-              <Tree
-                defaultExpandAll={true}
-                showLine={true}
-                switcherIcon={({expanded}) => (
-                  <CaretDownOutlined
-                    style={{
-                      transform: `rotate(${expanded ? 0 : -90}deg)`,
-                      transition: 'transform 0.3s',
-                    }}
-                  />
-                )}
-                onExpand={onExpand}
-                expandedKeys={expandedKeys}
-                autoExpandParent={autoExpandParent}
-                onSelect={(selectedPaths) => handleSelectElement(selectedPaths[0])}
-                selectedKeys={[selectedElement.path]}
-                treeData={treeData}
-                className={styles.sourceTree}
-              />
-            </div>
-          ) : (
-            <Tree treeData={[]} />
-          )}
-        </Spin>
-      </div>
-    </Card>
-  );
-};
-
-export default AppSource;
