+++ app/common/renderer/components/SessionInspector/SourceTab/AppSource/AppSource.module.css
